requirements.txt
ads==0.12.3
amqp==1.4.9
anyjson==0.3.3
apns2==0.7.2
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.3.0.23
captcha==0.3
catalogue==2.0.8
celery==3.1.18
certifi==2022.9.24
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
Django==3.2.16
django-admin-sortable2==0.6.18
django-allauth==0.43.0
django-appconf==1.0.2
django-authtools==2.0.0
django-bootstrap-v5==1.0.7
django-ckeditor==6.3.2
django-classy-tags==4.0.0
django-colorfield==0.1.14
django-compressor==4.1
django-countries==4.5
django-crispy-forms==1.6.1
django-cuser==2017.3.16
django-db-mailer==2.3.17
django-el-pagination==3.1.0
django-extensions==3.2.1
django-filter==22.1
django-fsm==2.4.0
django-heroku==0.3.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-jsonfield==1.0.1
django-newsletter==0.6
django-parsley==0.2a0
django-phone-field==1.8.1
django-push-notifications==1.6.0
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-rq==2.6.0
django-sekizai==0.10.0
django-simple-captcha==0.5.14
django-smart-selects==1.5.2
django-smtp-ssl==1.0
django-solo==2.0.0
django-static-jquery==2.1.4
django-tinymce==3.5.0
django-treebeard==4.1.0
django-utils-six==2.0
django-webtest==1.9.10
django-wkhtmltopdf==3.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-excel==2.2.0
drf-yasg==1.21.4
easy-thumbnails==2.4.1
et-xmlfile==1.1.0
filelock==3.8.2
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httpretty==0.8.10
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
jet==1.0
Jinja2==3.1.2
knpay-plugin==0.1.3
kombu==3.0.37
ldif3==3.1.1
MarkupSafe==2.1.1
mock==4.0.3
modeltranslation==0.25
mysqlclient==2.0.3
numpy==1.21.6
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
platformdirs==2.6.0
psycopg2==2.9.5
psycopg2-binary==2.8.5
py-vapid==1.9.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-card-me==0.9.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
pywebpush==1.14.0
PyYAML==6.0
raven==6.1.0
rcssmin==1.1.0
redis==2.10.3
render==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rq==1.11.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.3.1
sqlparse==0.4.3
surlex==0.2.0
typing_extensions==4.4.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.2.2
whitenoise==6.2.0
zipp==3.11.0
